Add methods for low-level image resource handling
With this, the WinForms hello-world is mostly usable, except for the
docking form, and assuming:

  * the assembly contains no binary resources, i.e., no BMPs, ICOs, etc.

  * a net9.0 or later SDK is used; this also reduces binary size by
    dozens of MiB

+++ lib/Npp.DotNet.Plugin/Winforms/Docking.cs
@@ -125,7 +125,7 @@ namespace Npp.DotNet.Plugin.Winforms
         /// <summary>
         /// Icon for tabs
         /// </summary>
-        public uint HIconTab;
+        public IntPtr HIconTab;
         /// <summary>
         /// Additional information (shown in window's title bar)
         /// </summary>
